qwre

+++ src/com/company/Main.java
@@ -18,7 +18,7 @@ public class Main {
         System.out.printf("Number1 * Number2=");
         System.out.println(num1*num2);
         System.out.printf("Number1 / Number2=");
-        System.out.println(num1/num2);
+        System.out.println((float)num1/num2);
 
 
     }
